requirements.txt
selenium
yaml
python-docx
qrcode
cairosvg